Chapter-04/website/src/requirements/dev.txt
-r _base.txt 
coverage 
django-debug-toolbar 
selenium